+++ packages/discord-bot/src/commands/praise.ts
@@ -1,10 +1,6 @@
-import {
-  SlashCommandBuilder,
-  SlashCommandSubcommandBuilder,
-} from '@discordjs/builders';
+import { SlashCommandBuilder } from '@discordjs/builders';
 import { APIMessage } from 'discord-api-types/v9';
 import logger from 'jet-logger';
-import { activationHandler } from '../handlers/activate';
 import { praiseHandler } from '../handlers/praise';
 import { Command } from '../interfaces/Command';
 import { getMsgLink } from '../utils/format';
@@ -12,32 +8,20 @@ import { getMsgLink } from '../utils/format';
 export const praise: Command = {
   data: new SlashCommandBuilder()
     .setName('praise')
-    .setDescription('Commands to interact with the praise system')
-    .addSubcommand(
-      new SlashCommandSubcommandBuilder()
-        .setName('dish')
-        .setDescription('Dish praise to a User')
-        .addStringOption((option) =>
-          option
-            .setName('receivers')
-            .setDescription(
-              'Mention the users you would like to send this praise to'
-            )
-            .setRequired(true)
-        )
-        .addStringOption((option) =>
-          option
-            .setName('reason')
-            .setDescription('The reason for this Praise')
-            .setRequired(true)
-        )
-    )
-    .addSubcommand(
-      new SlashCommandSubcommandBuilder()
-        .setName('activate')
+    .setDescription('Dish praise to a User')
+    .addStringOption((option) =>
+      option
+        .setName('receivers')
         .setDescription(
-          'Activate your praise account by linking your eth address'
+          'Mention the users you would like to send this praise to'
         )
+        .setRequired(true)
+    )
+    .addStringOption((option) =>
+      option
+        .setName('reason')
+        .setDescription('The reason for this Praise')
+        .setRequired(true)
     ),
 
   async execute(interaction) {
@@ -45,28 +29,18 @@ export const praise: Command = {
       if (!interaction.isCommand() || interaction.commandName !== 'praise')
         return;
 
-      const subCommand = interaction.options.getSubcommand();
-
-      switch (subCommand) {
-        case 'activate':
-          await activationHandler(interaction);
-          break;
-        case 'dish': {
-          const msg = (await interaction.deferReply({
-            fetchReply: true,
-          })) as APIMessage | void;
-          if (msg === undefined) return;
-          await praiseHandler(
-            interaction,
-            getMsgLink(
-              interaction.guildId || '',
-              interaction.channelId || '',
-              msg.id
-            )
-          );
-          break;
-        }
-      }
+      const msg = (await interaction.deferReply({
+        fetchReply: true,
+      })) as APIMessage | void;
+      if (msg === undefined) return;
+      await praiseHandler(
+        interaction,
+        getMsgLink(
+          interaction.guildId || '',
+          interaction.channelId || '',
+          msg.id
+        )
+      );
     } catch (err) {
       logger.err(err);
     }